refactor: Spring MVC 스타일로 폴더 및 모듈 아키텍처 재구성
